% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project_Code.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Model}
\usage{
my_lm(y, x, alpha, approach)
}
\arguments{
\item{y}{The \code{vector} dependent variable}

\item{x}{The \code{matrix} independent variable}

\item{alpha}{The \code{numeric} significance devlevel input by the user to estimate the confidence intervals}

\item{approach}{The \code{character} allows the user to choose bootstrap or asymptotic approach}
}
\value{
A \code{list} which returns the following attributes:
\describe{
     \item{ci.beta}{The confidence intervals}
     \item{sigma2.hat}{The standard deviation}
     \item{var.beta}{Variance}
     \item{MSPE}{The mean square prediction error}
     \item{Fstar}{An F Test}
     \item{Proboftest}{The probability or p-value of test}
}
}
\description{
Perform a linear regression based on a user input alpha value for the confidence intervals and
gives the user a choice beteen using the asymptotic or bootstrap approach.
The package returns the mean square prediction error (MSPE), an F-test, and creates plots for the residuals,
residual vs. fitted values, and a histogram of the residuals.
}
\examples{
my_lm (y= c(18, 45, 22), x=cbind(rbinom(n, size = 22), rnorm(n, 42, 65))), alpha = 0.01, approach =  bootstrap
my_lm (y= c(458, 32, 99) x=cbind(rbinom (n, size = 45), rnorm(n, 85, 22))), alpha = 0.10, approach = asymptotic
}
